+++ kinto/tests/test_main.py
@@ -27,7 +27,7 @@ class TestMain(unittest.TestCase):
             pass
 
     def test_cli_init_generates_configuration(self):
-        res = main(['--ini', TEMP_KINTO_INI, '--backend', 'memory', 'init'])
+        res = main(['--ini', TEMP_KINTO_INI, 'init', '--backend', 'memory'])
         assert res == 0
         assert os.path.exists(TEMP_KINTO_INI)
 
@@ -35,8 +35,8 @@ class TestMain(unittest.TestCase):
         with open(TEMP_KINTO_INI, 'w') as f:
             f.write("exists")
         with mock.patch('sys.stderr', new_callable=StringIO) as mock_stderr:
-            res = main(['--ini', TEMP_KINTO_INI,
-                        '--backend', 'memory', 'init'])
+            res = main(['--ini', TEMP_KINTO_INI, 'init',
+                        '--backend', 'memory'])
             assert res == 1
             content = mock_stderr.getvalue()
             assert '{} already exists.'.format(TEMP_KINTO_INI) in content
@@ -84,8 +84,8 @@ class TestMain(unittest.TestCase):
                     assert mocked_pip.call_count == 1
 
     def test_main_takes_sys_argv_by_default(self):
-        testargs = ["prog", "--ini", TEMP_KINTO_INI,
-                    '--backend', 'memory', 'init']
+        testargs = ["prog", "--ini", TEMP_KINTO_INI, 'init',
+                    '--backend', 'memory']
         with mock.patch.object(sys, 'argv', testargs):
             main()
 
@@ -95,17 +95,29 @@ class TestMain(unittest.TestCase):
 
     def test_cli_migrate_command_runs_init_schema(self):
         with mock.patch('kinto.__main__.scripts.migrate') as mocked_migrate:
-            res = main(['--ini', TEMP_KINTO_INI,
-                        '--backend', 'memory', 'init'])
+            res = main(['--ini', TEMP_KINTO_INI, 'init',
+                        '--backend', 'memory'])
             assert res == 0
             res = main(['--ini', TEMP_KINTO_INI, 'migrate'])
             assert res == 0
             assert mocked_migrate.call_count == 1
 
+    def test_cli_delete_collection_run_delete_collection_script(self):
+        with mock.patch('kinto.__main__.scripts.delete_collection') as del_col:
+            del_col.return_value = mock.sentinel.del_col_code
+            res = main(['--ini', TEMP_KINTO_INI, 'init',
+                        '--backend', 'memory'])
+            assert res == 0
+            res = main(['--ini', TEMP_KINTO_INI, 'delete-collection',
+                        '--bucket', 'test_bucket',
+                        '--collection', 'test_collection'])
+            assert res == mock.sentinel.del_col_code
+            assert del_col.call_count == 1
+
     def test_cli_start_runs_pserve(self):
         with mock.patch('kinto.__main__.pserve.main') as mocked_pserve:
-            res = main(['--ini', TEMP_KINTO_INI,
-                        '--backend', 'memory', 'init'])
+            res = main(['--ini', TEMP_KINTO_INI, 'init',
+                        '--backend', 'memory'])
             assert res == 0
             res = main(['--ini', TEMP_KINTO_INI, 'start'])
             assert res == 0
@@ -113,8 +125,8 @@ class TestMain(unittest.TestCase):
 
     def test_cli_start_with_reload_runs_pserve_with_reload(self):
         with mock.patch('kinto.__main__.pserve.main') as mocked_pserve:
-            res = main(['--ini', TEMP_KINTO_INI,
-                        '--backend', 'memory', 'init'])
+            res = main(['--ini', TEMP_KINTO_INI, 'init',
+                        '--backend', 'memory'])
             assert res == 0
             res = main(['--ini', TEMP_KINTO_INI, 'start', '--reload'])
             assert res == 0